requirements.txt
google-adk[extensions]>=1.27.0
google-genai>=0.8.0
litellm>=1.61.0
openai>=1.76.0
supabase>=2.10.0
python-dotenv>=1.0.1
pydantic>=2.9.0
httpx>=0.27.0
psycopg[binary]>=3.2.0
